Remove use of float/double.
Many embedded platforms don't support floating point, so avoid using
them. Delegate handling of floating point to WASM runners.

+++ src/wasmjit/ast.h
@@ -365,10 +365,10 @@ struct Instr {
 			uint64_t value;
 		} i64_const;
 		struct {
-			float value;
+			uint32_t value;
 		} f32_const;
 		struct {
-			double value;
+			uint64_t value;
 		} f64_const;
 	} data;
 };
